Fix compilation errors in 15+ crypto source files: type conversions, sign-conversions, unused variables, parameter shadowing
- sm3.cpp: Fixed loop variables from int to size_t in W expansion (8 errors)
- dsa.cpp: Added NumBytes() casts to size_t for NTL integration (7 errors)
- ffs1.cpp: Fixed parameter shadowing (min/max), array index conversions (7 errors)
- otp.cpp: Fixed type conversions in HMAC and base32 operations (7 errors)
- ecdh.cpp: Fixed get_bit_size() conversions and bit shift casts (5 errors)
- dh.cpp: Added NumBytes() casts for key export functions (2 errors)
- ecdsa.cpp: Fixed get_bit_size() and NumBytes() conversions (3 errors)
- zuc.cpp: Changed loop variable from int to uint32_t (2 errors)
- ffsa.cpp: Cast int8_t to size_t for vector reserve() (2 errors)
- zkp.cpp: Cast size_t to long for ZZ constructor calls (2 errors)
- ecies.cpp: Cast bit XOR result to uint8_t (1 error)
- sm2.cpp: Removed unused curve variable (1 error)
- ecc_curve.cpp: Fixed (bit_size+7)/8 division casts (1 error)
- rsa.cpp: Fixed multiple (bits+7)/8 conversions and bit shift operations
- simd.cpp: Fixed uint32_t to int32_t casts for SIMD intrinsics
- ecdh.cpp, ecies.cpp: Additional sign-conversion fixes

Total: 33 lower-priority compilation errors fixed across 15 files
Remaining issues: sha256.cpp, aes.cpp SHA256 SIMD intrinsics and external GoogleTest library warnings

+++ src/advanced/zk/ffs/ffsa.cpp
@@ -102,7 +102,7 @@ void ffsa_get_modulus(mpz_t result)
 bool_vec_t ffsa_get_bool_vector(int8_t k)
 {
     bool_vec_t a;
-    a.reserve(k);
+    a.reserve(static_cast<size_t>(k));
 
     srand(static_cast<unsigned int>(time(NULL)));
 
@@ -121,7 +121,7 @@ bool_vec_t ffsa_get_bool_vector(int8_t k)
 void ffsa_get_secrets(mpz_vec_t& s, int8_t k, const mpz_t n)
 {
     s.clear();
-    s.reserve(k);
+    s.reserve(static_cast<size_t>(k));
 
     gmp_randstate_t state;
     gmp_randinit_default(state);
