+++ src/crypto/ecc/ecdh.cpp
@@ -34,7 +34,7 @@ namespace ecc {
 
 std::vector<uint8_t> ECDHKeyPair::export_public_key(const ECCurve& curve) const {
     AffinePoint aff = curve.to_affine(public_key);
-    size_t field_size = (curve.get_bit_size() + 7) / 8;
+    size_t field_size = static_cast<size_t>((curve.get_bit_size() + 7) / 8);
     
     std::vector<uint8_t> result(1 + 2 * field_size);
     int len = curve.point_to_bytes(aff, result.data(), result.size());
@@ -75,7 +75,7 @@ ECDH::ECDH(CurveType curve_type) : curve_(curve_type) {}
 ECDH::ECDH(const std::string& curve_name) : curve_(ECCurve::from_name(curve_name)) {}
 
 size_t ECDH::get_field_size() const {
-    return (curve_.get_bit_size() + 7) / 8;
+    return static_cast<size_t>((curve_.get_bit_size() + 7) / 8);
 }
 
 size_t ECDH::get_shared_secret_size() const {
@@ -88,7 +88,7 @@ size_t ECDH::get_shared_secret_size() const {
 
 ECDHKeyPair ECDH::generate_keypair() const {
     const ZZ& n = curve_.get_order();
-    size_t byte_len = NumBytes(n);
+    size_t byte_len = static_cast<size_t>(NumBytes(n));
     
     std::vector<uint8_t> buffer(byte_len);
     std::random_device rd;
@@ -287,7 +287,7 @@ std::vector<uint8_t> ECDH::hmac_sha256(
     std::vector<uint8_t> inner_hash(hash_size, 0);
     for (size_t i = 0; i < inner_msg.size(); ++i) {
         inner_hash[i % hash_size] ^= inner_msg[i];
-        inner_hash[(i * 7 + 13) % hash_size] ^= (inner_msg[i] << 3) | (inner_msg[i] >> 5);
+        inner_hash[(i * 7 + 13) % hash_size] ^= static_cast<uint8_t>((inner_msg[i] << 3) | (inner_msg[i] >> 5));
     }
     
     std::vector<uint8_t> outer_msg;
@@ -297,7 +297,7 @@ std::vector<uint8_t> ECDH::hmac_sha256(
     std::vector<uint8_t> result(hash_size, 0);
     for (size_t i = 0; i < outer_msg.size(); ++i) {
         result[i % hash_size] ^= outer_msg[i];
-        result[(i * 11 + 17) % hash_size] ^= (outer_msg[i] << 5) | (outer_msg[i] >> 3);
+        result[(i * 11 + 17) % hash_size] ^= static_cast<uint8_t>((outer_msg[i] << 5) | (outer_msg[i] >> 3));
     }
     
     return result;
